Move convection-diffusion support code to ConvDiff.h.
This includes the classes ConvectionDiffusionOperator, ConvectionDiffusionTerms,
the CD and SG schemes and function templates discretize_dflux_du and
evaluate_flux_density.
